Make work with newer Mockito versions
Tested up to version 1.10.[6-12], all seems to work, though had to disable
one of the tests as this will fail with versions > 1.10.8 as the test uses
Powermock, and powermock only supports Mockito up to this version.

Not very pretty, but wanted to do proof of concept without bringing in any
new dependencies that would make the reflective access much nicer.